% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ep_haplotype_name.R
\name{check_ep_haplotype_name}
\alias{check_ep_haplotype_name}
\title{Internal function}
\usage{
check_ep_haplotype_name(ep_haplotype_name)
}
\arguments{
\item{ep_haplotype_name}{a haplotype name as used internally
by EpitopePrediction,
Use \link{to_ep_haplotype_name} to convert a formal haplotype
name to an EpitopePrediction haplotype name}
}
\description{
Internal function to check if the EpitopePrediction name is valid.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
