% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_haplotype_name.R
\name{check_haplotype_name}
\alias{check_haplotype_name}
\title{Check if the haplotype name is valid and supported}
\usage{
check_haplotype_name(haplotype_name)
}
\arguments{
\item{haplotype_name}{a formal haplotype name,
Use \link{to_ep_haplotype_name} to convert this to an
EpitopePrediction haplotype name}
}
\description{
Check if the haplotype name is valid and supported
by EpitopePrediction.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
