% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_protein_sequence.R
\name{check_protein_sequence}
\alias{check_protein_sequence}
\title{Internal function}
\usage{
check_protein_sequence(protein_sequence)
}
\arguments{
\item{protein_sequence}{a protein sequence}
}
\description{
Internal function to check the protein sequence.
}
\examples{
check_protein_sequence("FAMILYVW")
}
\author{
Richèl J.C. Bilderbeek
}
