% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_haplotype_ids_to_haplotypes.R
\name{convert_haplotype_ids_to_haplotypes}
\alias{convert_haplotype_ids_to_haplotypes}
\title{Internal function}
\usage{
convert_haplotype_ids_to_haplotypes(haplotype_ids)
}
\arguments{
\item{haplotype_ids}{one or more \code{epiprepreds} haplotype IDs,
for example \code{h1}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{epiprepreds} haplotype ID to an EpitopePrediction haplotype name}
}
\description{
Internal function to convert one or more haplotype IDs
to EpitopePrediction haplotype names
}
\examples{
convert_haplotype_ids_to_haplotypes("h13")
}
\author{
Richèl J.C. Bilderbeek
}
