% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_haplotype_to_haplotype_id.R
\name{convert_haplotype_to_haplotype_id}
\alias{convert_haplotype_to_haplotype_id}
\title{Internal function}
\usage{
convert_haplotype_to_haplotype_id(ep_haplotype_name)
}
\arguments{
\item{ep_haplotype_name}{a haplotype name as used internally
by EpitopePrediction,
Use \link{to_ep_haplotype_name} to convert a formal haplotype
name to an EpitopePrediction haplotype name}
}
\description{
Internal function to convert an EpitopePrediction haplotype name
to a hapolotype ID
}
\examples{
convert_haplotype_to_haplotype_id("HLA-A-33:01")
}
\author{
Richèl J.C. Bilderbeek
}
