% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_haplotypes_to_haplotype_ids.R
\name{convert_haplotypes_to_haplotype_ids}
\alias{convert_haplotypes_to_haplotype_ids}
\title{Internal function}
\usage{
convert_haplotypes_to_haplotype_ids(ep_haplotype_names)
}
\arguments{
\item{ep_haplotype_names}{one or more haplotype name
as used internelly by EpitopePrediction,
Use \link{to_ep_haplotype_name} to convert a formal haplotype
name to an EpitopePrediction haplotype name}
}
\description{
Internal function to convert one or more
EpitopePrediction haplotype names
to haplotype IDs
}
\examples{
convert_haplotypes_to_haplotype_ids("HLA-A-33:01")
}
\author{
Richèl J.C. Bilderbeek
}
