% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  binding_strength_threshold,
  ep_haplotype_name,
  ep_haplotype_names,
  fasta_filename,
  haplotype,
  haplotype_name,
  haplotypes,
  haplotype_id,
  haplotype_ids,
  ic50_threshold,
  max_hydrophobicity,
  mhc_haplotype,
  min_hydrophobicity,
  n_peptides,
  peptide_length,
  protein_sequence,
  protein_sequences,
  sink_filename,
  verbose
)
}
\arguments{
\item{binding_strength_threshold}{threshold of the MHC2 allele's
binding strength. A low threshold, e.g. \code{5.0}, selects for the top
5 percent best binders}

\item{ep_haplotype_name}{a haplotype name as used internally
by EpitopePrediction,
Use \link{to_ep_haplotype_name} to convert a formal haplotype
name to an EpitopePrediction haplotype name}

\item{ep_haplotype_names}{one or more haplotype name
as used internelly by EpitopePrediction,
Use \link{to_ep_haplotype_name} to convert a formal haplotype
name to an EpitopePrediction haplotype name}

\item{fasta_filename}{name of a FASTA file}

\item{haplotype}{the EpitopePrediction haplotype name,
for example \code{HLA-A33:01}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{epiprepreds} haplotype ID to an EpitopePrediction haplotype name}

\item{haplotype_name}{a formal haplotype name,
Use \link{to_ep_haplotype_name} to convert this to an
EpitopePrediction haplotype name}

\item{haplotypes}{one ore more EpitopePrediction haplotype IDs,
for example \code{HLA-A33:01}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{epiprepreds} haplotype ID to an EpitopePrediction haplotype name}

\item{haplotype_id}{the \code{epiprepreds} haplotype ID,
for example \code{h1}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{epiprepreds} haplotype ID to an EpitopePrediction haplotype name}

\item{haplotype_ids}{one or more \code{epiprepreds} haplotype IDs,
for example \code{h1}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{epiprepreds} haplotype ID to an EpitopePrediction haplotype name}

\item{ic50_threshold}{peptides with a predicted IC50 value (in nM)
lower than this will be considered binders.
A threshold of 500 (nM) is common.}

\item{max_hydrophobicity}{the maximal hydrophobicity}

\item{mhc_haplotype}{haplotype of either HMC-I or MHC-II.
in EpitopePredictionr format.
Use \code{EpitopePredictionr::get_mhc_1_haplotypes}
to get a list of all MHC-I haplotypes.
Use \code{EpitopePredictionr::get_mhc_2_haplotypes}
to get a list of all MHC-II haplotypes.}

\item{min_hydrophobicity}{the minimal hydrophobicity}

\item{n_peptides}{number of peptides}

\item{peptide_length}{length of the peptide in amino acids}

\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}

\item{sink_filename}{temporary filename to \link{sink} output to.
This is \code{/dev/null} by default, which works on Linux perfectly}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
