% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ep_haplotype_names.R
\name{get_ep_haplotype_names}
\alias{get_ep_haplotype_names}
\title{Get the EpitopePrediction haplotype names}
\usage{
get_ep_haplotype_names(sink_filename = "/dev/null")
}
\arguments{
\item{sink_filename}{temporary filename to \link{sink} output to.
This is \code{/dev/null} by default, which works on Linux perfectly}
}
\description{
Get the EpitopePrediction haplotype names
}
\examples{
get_ep_haplotype_names()
}
