% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_haplotype_lut.R
\name{get_haplotype_lut}
\alias{get_haplotype_lut}
\title{Internal function}
\usage{
get_haplotype_lut()
}
\description{
Create the mapping from the EpitopePrediction haplotype names
to a haplotype ID
}
\examples{
get_haplotype_lut()
}
