% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ic50_threshold.R
\name{get_ic50_threshold}
\alias{get_ic50_threshold}
\title{Look up the IC50 value of a certain quantile}
\usage{
get_ic50_threshold(peptide_length = 9, haplotype_name, percentile = 0.02)
}
\arguments{
\item{peptide_length}{length of the peptide in amino acids}

\item{haplotype_name}{a formal haplotype name,
Use \link{to_ep_haplotype_name} to convert this to an
EpitopePrediction haplotype name}

\item{percentile}{the IC50 percentile. The IC50 that separates the top
\code{percentile} from the worse binders}
}
\description{
Look up the IC50 value of a certain quantile.
Such a value can serve as a threshold:
for example, to get the IC50 value from below which the
peptides bind strongest.
}
