% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lut_filename.R
\name{get_lut_filename}
\alias{get_lut_filename}
\title{Internal function}
\usage{
get_lut_filename(peptide_length, ep_haplotype_name)
}
\arguments{
\item{peptide_length}{length of the peptide in amino acids}

\item{ep_haplotype_name}{a haplotype name as used internally
by EpitopePrediction,
Use \link{to_ep_haplotype_name} to convert a formal haplotype
name to an EpitopePrediction haplotype name}
}
\description{
Internal function to get the filename for a lookup-table
for a specific peptide size and EpitopePrediction haplotype.
}
