% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ep_haplotype_name.R
\name{is_ep_haplotype_name}
\alias{is_ep_haplotype_name}
\title{Is the haplotype name an EpitopePrediction haplotype name?}
\usage{
is_ep_haplotype_name(ep_haplotype_name)
}
\arguments{
\item{ep_haplotype_name}{a haplotype name as used internally
by EpitopePrediction,
Use \link{to_ep_haplotype_name} to convert a formal haplotype
name to an EpitopePrediction haplotype name}
}
\value{
TRUE if the haplotype name is a valid
  EpitopePrediction haplotype name, FALSE otherwise
}
\description{
Is the haplotype name an EpitopePrediction haplotype name?
}
