% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_epipred_haplotype.R
\name{is_epipred_haplotype}
\alias{is_epipred_haplotype}
\title{Is the haplotype a valid EpitopePrediction hapolotype?}
\usage{
is_epipred_haplotype(haplotype)
}
\arguments{
\item{haplotype}{the EpitopePrediction haplotype ID,
for example \code{HLA-A33:01}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{epiprepreds} haplotype ID to an EpitopePrediction haplotype name}
}
\description{
Is the haplotype a valid EpitopePrediction hapolotype?
}
