% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_haplotype_name.R
\name{is_haplotype_name}
\alias{is_haplotype_name}
\title{Is the haplotype name valid and supported?}
\usage{
is_haplotype_name(haplotype_name)
}
\arguments{
\item{haplotype_name}{a formal haplotype name,
Use \link{to_ep_haplotype_name} to convert this to an
EpitopePrediction haplotype name}
}
\value{
TRUE if the haplotype name is a valid and supported
  haplotype name, FALSE otherwise
}
\description{
Is the haplotype name a valid and supported haplotype?
}
