% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_protein_sequence.R
\name{is_protein_sequence}
\alias{is_protein_sequence}
\title{Internal function}
\usage{
is_protein_sequence(protein_sequence)
}
\arguments{
\item{protein_sequence}{a protein sequence}
}
\value{
TRUE if the protein sequence is valid
}
\description{
Internal function to determine if a the protein sequence is valid.
}
\examples{
is_protein_sequence("FAMILYVW")
}
\author{
Richèl J.C. Bilderbeek
}
