% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ic50s.R
\name{predict_ic50s}
\alias{predict_ic50s}
\title{Predict IC50s}
\usage{
predict_ic50s(protein_sequence, peptide_length, haplotype_name)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{peptide_length}{length of the peptide in amino acids}

\item{haplotype_name}{a formal haplotype name,
Use \link{to_ep_haplotype_name} to convert this to an
EpitopePrediction haplotype name}
}
\description{
Predict IC50s
}
\author{
Richèl J.C. Bilderbeek
}
