% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shred_protein.R
\name{shred_protein}
\alias{shred_protein}
\title{Internal function}
\usage{
shred_protein(protein_sequence, peptide_length)
}
\arguments{
\item{protein_sequence}{a protein sequence}

\item{peptide_length}{length of the peptide in amino acids}
}
\description{
Internal function to 'shred a protein', that is,
to obtain all peptides of length \code{peptide_length}
from the protein sequence
}
\examples{
shred_protein(
  protein_sequence = "FANTASTICALLY",
  peptide_length = 9
)
}
\author{
Richèl J.C. Bilderbeek
}
