% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ep_haplotype_name.R
\name{to_ep_haplotype_name}
\alias{to_ep_haplotype_name}
\title{Convert a haplotype name to an EpitopePrediction name}
\usage{
to_ep_haplotype_name(haplotype_name)
}
\arguments{
\item{haplotype_name}{a formal haplotype name,
Use \link{to_ep_haplotype_name} to convert this to an
EpitopePrediction haplotype name}
}
\description{
Convert a formally named (MHC-I) haplotype name
to one to be used by the EpitopePrediction package
}
\examples{
to_ep_haplotype_name("HLA-A*01:01")
}
\author{
Richèl J.C. Bilderbeek
}
