% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_epipred_name.R
\name{to_epipred_name}
\alias{to_epipred_name}
\title{Convert to EpitopePrediction name}
\usage{
to_epipred_name(haplotype)
}
\description{
Convert a formally named MHC-I haplotype
to its EpitopePrediction counterpart.
}
