% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_haplotype_name.R
\name{check_haplotype_name}
\alias{check_haplotype_name}
\title{Check if the haplotype name is valid and supported}
\usage{
check_haplotype_name(haplotype_name, sink_filename = "/dev/null")
}
\arguments{
\item{haplotype_name}{a formal haplotype name,
Use \link{to_ep_haplotype_name} to convert this to an
EpitopePrediction haplotype name}

\item{sink_filename}{temporary filename to \link{sink} output to.
This is \code{/dev/null} by default, which works on Linux perfectly}
}
\description{
Check if the haplotype name is valid and supported
by EpitopePrediction.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
