% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_haplotype_to_haplotype_id.R
\name{convert_haplotype_to_haplotype_id}
\alias{convert_haplotype_to_haplotype_id}
\title{Convert a haplotype to a haplotype ID}
\usage{
convert_haplotype_to_haplotype_id(haplotype)
}
\arguments{
\item{haplotype}{the EpitopePrediction haplotype ID,
for example \code{HLA-A33:01}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{epiprepreds} haplotype ID to an EpitopePrediction haplotype name}
}
\description{
Convert a haplotype (in EpitopePrediction naming format) to a hapolotype ID
}
\examples{
convert_haplotype_to_haplotype_id("HLA-A-33:01")
}
\author{
Richèl J.C. Bilderbeek
}
