% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_epipred_name.R
\name{to_epipred_name}
\alias{to_epipred_name}
\title{Convert to EpitopePrediction name}
\usage{
to_epipred_name(haplotype)
}
\arguments{
\item{haplotype}{the EpitopePrediction haplotype ID,
for example \code{HLA-A33:01}.
Use \link{get_haplotype_lut} to get a lookup table that maps
a \code{epiprepreds} haplotype ID to an EpitopePrediction haplotype name}
}
\description{
Convert a formally named MHC-I haplotype
to its EpitopePrediction counterpart.
}
